// projeto/web/vite_react_p2/src/services/generated/models/createUserDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */

export interface CreateUserDto {
  /** @nullable */
  email: string | null;
  /**
   * @minLength 8
   * @nullable
   * @pattern ^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[\W_]).+$
   */
  password: string | null;
  /**
   * @minLength 8
   * @nullable
   * @pattern ^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[\W_]).+$
   */
  confirmPassword: string | null;
  /**
   * @minLength 3
   * @maxLength 255
   * @nullable
   */
  name: string | null;
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  roleId: number | string | null;
}