// projeto/web/vite_react_p2/src/services/generated/models/responseUserDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */
import type { EUserStatus } from './eUserStatus';
import type { ResponseRoleDto } from './responseRoleDto';

export interface ResponseUserDto {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id?: number | string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  name?: string | null;
  status?: EUserStatus;
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  roleId?: number | string | null;
  role?: ResponseRoleDto;
}